requirements.txt
requests
pandas
openpyxl
python-dotenv
pytest
pytest-mock
tabulate